Move 3D Viewer to applications; set 3d_viewer as default app
