+++ app/3dViewer/src/main/java/org/phoebus/applications/viewer3d/InstallExamplesMenuEntry.java
@@ -5,30 +5,30 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.phoebus.app.viewer3d;
+package org.phoebus.applications.viewer3d;
 
 import org.phoebus.framework.spi.MenuEntry;
-import org.phoebus.framework.workbench.ApplicationService;
+import org.phoebus.ui.examples.ExampleInstaller;
 import org.phoebus.ui.javafx.ImageCache;
 
 import javafx.scene.image.Image;
 
 /** Menu entry
- *  @author Evan Smith
+ *  @author Kay Kasemir
  */
 @SuppressWarnings("nls")
-public class Viewer3dMenuEntry implements MenuEntry
+public class InstallExamplesMenuEntry implements MenuEntry
 {
     @Override
     public String getName()
     {
-        return Viewer3dApp.DISPLAY_NAME;
+        return "Install Example 3D Viewer Files";
     }
 
     @Override
     public String getMenuPath()
     {
-        return "Display";
+        return "Display.Examples";
     }
 
     @Override
@@ -40,7 +40,14 @@ public class Viewer3dMenuEntry implements MenuEntry
     @Override
     public Void call() throws Exception
     {
-        ApplicationService.createInstance(Viewer3dApp.NAME);
-        return null;
+        new ExampleInstaller("Select Directory for Installing 3D Viewer Examples",
+                Viewer3dApp.class.getResource("/3d_viewer_examples"),
+                "3D Viewer",
+                "NaCl.shp",
+                Viewer3dApp.NAME,
+                "Examples have been installed in\n\n  {0}")
+           .call();
+
+           return null;
     }
 }
