+++ app/3dViewer/src/main/java/org/phoebus/applications/viewer3d/Preferences.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.phoebus.app.viewer3d;
+package org.phoebus.applications.viewer3d;
 
 import org.phoebus.framework.preferences.PreferencesReader;
