+++ app/3dViewer/src/main/resources/META-INF/services/org.phoebus.framework.spi.AppResourceDescriptor
@@ -1,1 +1,1 @@
-org.phoebus.app.viewer3d.Viewer3dApp
+org.phoebus.applications.viewer3d.Viewer3dApp
