+++ core/framework/src/main/resources/workbench_preferences.properties
@@ -19,7 +19,7 @@
 # Example:
 #
 # Start 'gedit' for text files, 'eog' for images, 'firefox' for PDF files
-# external_apps=Text Editor,txt|dat|py|ini|db|xml|xsl|css|cmd|sh|st|log|out|md,gedit; Image Viewer,png|jpg|gif|jpeg,eog; PDF Viewer,pdf,firefox
+# external_apps=Text Editor,txt|dat|py|ini|db|xml|xsl|css|cmd|sh|st|log|out|md|shp,gedit; Image Viewer,png|jpg|gif|jpeg,eog; PDF Viewer,pdf,firefox
 
 
 #
