+++ core/ui/src/main/resources/phoebus_ui_preferences.properties
@@ -15,7 +15,10 @@ splash=true
 # For example, "run, exe" would pick "display_runtime" over "display_editor",
 # and "foo_executor" over "foo_creator".
 # The patterns "edit, creat" would inversely open the editor-type apps.
-default_apps=run
+#
+# This makes the display_runtime and the 3d_viewer default apps,
+# using display_editor and a potentially configured text editor for *.shp files secondary
+default_apps=run,3d
 
 # Top resources to show in "File" menu and toolbar
 #
