+++ phoebus-product/src/main/resources/logging.properties
@@ -49,6 +49,7 @@ org.phoebus.framework.rdb.level = WARNING
 org.phoebus.pv.level = WARNING
 org.phoebus.applications.filebrowser.level = WARNING
 org.phoebus.applications.pvtree.level = WARNING
+org.phoebus.applications.viewer3d.level = WARNING
 org.phoebus.archive.reader.level = WARNING
 org.phoebus.security.level = WARNING
 org.csstudio.trends.databrowser3.level = WARNING
